HZN-1455: initial attempt at support for running under Java 9
This commit requires the changes made in `jira/NMS-10518`, which
make it easier to manipulate the command-line options.

* conditionally apply `-Djava.endorsed.dirs=` to Java 8 only
* use `--add-modules=` on the command-line on Java 9+; the list
  of modules comes with a run of `jdeps` over the jars in a built
  OpenNMS install
* add `$OPENNMS_HOME/lib/endorsed` to the regular classpath in
  org.opennms.bootstrap.Bootstrap so that we can leave that
  directory for Java 8 backwards-compatibility
* update Jetty to the latest 9.4 to avoid module access issues in
  previous versions; I had to add a dependency to the `jspc` build
  in `opennms-webapp` to make this happy

Note: not everything works.

The PostgreSQL JDBC driver had to be updated and this introduced
an incompatibility with our existing Liquibase (and our Liquibase
is hard to upgrade for various esoteric reasons).  Currently
upgrading databases that were created with the previous JDBC
driver fails because it expects the `locked` column to be
upper-case and case-sensitive.

Also the XmpCollector and XmpDetector rely on an external
dependency (`org.krupczak.xmp.SocketOpts`) that throws an
exception from `sun.security.x509.DNSName`.

There are probably other things I'm missing as well.